Fix issue #4202. Move DTO to gitrepo-api. Change gerrit-address
