+++ components/gitrepo-api/src/main/java/io/fabric8/repo/git/GerritGitApi.java
@@ -1,0 +1,13 @@
+package io.fabric8.repo.git;
+
+import javax.ws.rs.*;
+
+@Path("a")
+@Produces("application/json")
+@Consumes("application/json")
+public interface GerritGitApi {
+
+    @POST
+    @Path("projects/{repo}")
+    public GerritRepositoryDTO createRepository(@PathParam("repo") String repo, CreateRepositoryDTO dto);
+}
